When in dual boot mode, don't install grub onto the RP.

+++ ubiquity/dell-bootstrap.py
@@ -70,6 +70,9 @@ class PageNoninteractive(PluginUI):
     def show_reboot_dialog(self):
         pass
 
+    def show_dual_dialog(self):
+        pass
+
     def show_exception_dialog(self,e):
         pass
 
@@ -108,6 +111,8 @@ class PageGtk(PluginUI):
             self.hidden_radio = builder.get_object('hidden_radio')
             self.reboot_dialog = builder.get_object('reboot_dialog')
             self.reboot_dialog.set_title('Dell Recovery')
+            self.dual_dialog = builder.get_object('dual_dialog')
+            self.dual_dialog.set_title('Dell Recovery')
             self.info_box = builder.get_object('info_box')
             self.info_spinner = builder.get_object('info_spinner')
             self.err_dialog = builder.get_object('err_dialog')
@@ -223,6 +228,11 @@ class PageGtk(PluginUI):
         self.controller.toggle_top_level()
         self.info_spinner.stop()
         self.reboot_dialog.run()
+
+    def show_dual_dialog(self):
+        self.controller.toggle_top_level()
+        self.info_spinner.stop()
+        self.dual_dialog.run()
 
     def show_exception_dialog(self, e):
         self.info_spinner.stop()
@@ -419,7 +429,10 @@ class Page(Plugin):
         bus = dbus.SystemBus()
         bus.add_signal_receiver(reboot_machine, 'DeviceRemoved', 'org.freedesktop.UDisks')
 
-        self.ui.show_reboot_dialog()
+        if self.dual:
+            self.ui.show_dual_dialog()
+        else:
+            self.ui.show_reboot_dialog()
 
         reboot_machine(None)
 
@@ -752,18 +765,21 @@ class rp_builder(Thread):
         with misc.raised_privileges():
             magic.white_tree("copy", white_pattern, CDROM_MOUNT, '/boot')
 
-        #Check for a grub.cfg - replace as necessary
-        if os.path.exists(os.path.join('/boot', 'grub', 'grub.cfg')):
-            os.remove(os.path.join('/boot', 'grub', 'grub.cfg'))
-        with misc.raised_privileges():
-            magic.process_conf_file('/usr/share/dell/grub/recovery_partition.cfg', \
-                                    os.path.join('/boot', 'grub', 'grub.cfg'),     \
-                                    RP_PART, self.dual)
-
-        #Install grub
-        grub = misc.execute_root('grub-install', '--force', self.device + RP_PART)
-        if grub is False:
-            raise RuntimeError, ("Error installing grub to %s%s" % (self.device, RP_PART))
+        #Only prepare grub if this won't be a dual boot
+        if not self.dual:
+            #Check for a grub.cfg - replace as necessary
+            if os.path.exists(os.path.join('/boot', 'grub', 'grub.cfg')):
+                with misc.raised_privileges():
+                    os.remove(os.path.join('/boot', 'grub', 'grub.cfg'))
+            with misc.raised_privileges():
+                magic.process_conf_file('/usr/share/dell/grub/recovery_partition.cfg', \
+                                        os.path.join('/boot', 'grub', 'grub.cfg'),     \
+                                        RP_PART, self.dual)
+    
+            #Install grub (or configure our boot setup)
+            grub = misc.execute_root('grub-install', '--force', self.device + RP_PART)
+            if grub is False:
+                raise RuntimeError, ("Error installing grub to %s%s" % (self.device, RP_PART))
 
         #Build new UUID
         if int(self.mem) >= 1000000:
